Replace Rocket with Axum.
Axum gives us more control over the flow of requests and allows us to
use all the tower ecosystem, which has a lot of nice middlewares.

The first immediate benefit we get is better logging using the tracing
ecosystem and support for a request ID. Clients can send a
`x-request-id` header, which will be included in all log messages
printed by the request. If no ID is provided one is generated randomly
by the server. The ID, generated or provided by the client, is also
included in the response.

Most of our business code and endpoints are untouched, since Rocket and
Axum use fairly similar paradigms anyway. Most of the changes are in
auxilary code, such as testing, metrics and response formatting.