Add a custom Python README.
By default Hatch automatically includes the README.md file at the root
of the project as the Python package's description, which shows up when
looking at the package on PyPI. Since the README mostly contains
information about the server this could lead to confusion, with users
trying to start a server when what they really wanted was to be
redirected to the main outpack-py package.

This adds a minimal README with the Python specific information, and a
link to the main package. The new file is referenced from
`pyproject.toml`, ensuring this is what gets used as the description
rather than the default one.