Add an outpack_server_build_info metric.
This describes the server's version and build environment.
For now it only contains the version number (as specified in the
`Cargo.toml`), the git revision and the Rust version used in the build.

The versions are listed as labels, and the metric has a value of 1.
This follows the convention that is widespread in the Prometheus
ecosystem.

This is what the resulting metric might look like:
```
outpack_server_build_info{revision="dd6b97db1f52ece5ea4b635ec43655d005ad7627",rustc="1.78.0",version="0.3.0"} 1
```