Fix the outpack_server build.
GitHub has decomissioned upload-artifact v3 and is forcing everyone to
use v4. These work in similar ways, except v4 doesn't allow multiple
jobs to push to the same artifact name anymore. We need to use distinct
names for each OS/architecture pair, and have the call to
download-artifact pull and merge them all. This is based on the
[cibuildwheel documentation][cibuildwheel].

A couple of clippy fixes along the way.

[cibuildwheel]: https://cibuildwheel.pypa.io/en/stable/deliver-to-pypi/