Report an error when pushing conflicting metadata.
Our packet IDs are generally assumed to be globally unique, but this is
not guaranteed by construction: the ID is not a hash of its contents,
and the metadata store is not content-addressed.

This means one could try to upload two different packets to the store
with the same ID. Previously, outpack server would accept the second
request but skip writing to the file. It is better to detect this
situation and to return an HTTP error when it occurs.