Improve Docker container definition.
This changes the container from blocking on start until it finds a root
to creating it itself instead. This allows the service to be run out of
the box with just a `docker run outpack_server`.

In the past we were going to use a Git clone as the outpack root so it
made sense to have more control over the creation of it, but we've gone
down a different route now.

Additionally it adds [tini][tini] as an entrypoint wrapper to make sure
we have a well behaved PID 1. This avoids the issue of the container
taking a long time to shutdown.

[tini]: https://github.com/krallin/tini